ACL_PyTorch/contrib/cv/detection/RFCN/requirements.txt
cython
cffi
opencv-python
scipy
easydict
matplotlib
pyyaml
imageio
numpy
future
pytorch==1.7.0
torchvision==0.8.0
torchaudio==0.7.0